test proto::Any json ser-der
