+++ ibc-testkit/src/testapp/ibc/clients/mod.rs
@@ -126,22 +126,3 @@ impl TryFrom<AnyConsensusState> for MockConsensusState {
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use ibc::primitives::proto::Any;
-
-    use super::mock::header::MockHeader;
-    use crate::testapp::ibc::clients::mock::client_state::MockClientState;
-
-    #[test]
-    fn test_any_client_state_to_json() {
-        let client_state = MockClientState::new(MockHeader::default());
-        let expected = r#"{"typeUrl":"/ibc.mock.ClientState","value":"CgQKAhABEIDIr6Al"}"#;
-        let json = serde_json::to_string(&Any::from(client_state)).unwrap();
-        assert_eq!(json, expected);
-
-        let proto_any = serde_json::from_str::<Any>(expected).unwrap();
-        assert_eq!(proto_any, Any::from(client_state));
-    }
-}
